requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-dotenv==1.0.0
tensorflow==2.14.0
torch==2.1.1
transformers==4.35.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
aiohttp==3.9.1
asyncio==3.4.3
aiosqlite==0.19.0 